requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
anthropic
pydantic
